==> code/1 Naomi's working directory.R <==
#working directory here
#original code by A. R. Martinig
#last edited June 25, 2024 by N. Brunner

#Delete previous information stored 
rm(list=ls(all=T))

getwd() #this will give you your CURRENT working directory, you need to copy and paste it below and then add the extra bits that are needed to get you to the folder with the data

#output will be something like this:
#[1] "C:/Users/Programmar/Documents" 
#C:/Users/Programmar/Documents this is what you put below plus the extra stuff

##set wd to the folder with all your csv's in it
setwd("C:/Users/Owner/Documents/Gestation 2/gestation/data")
